Add double

+++ CalcArrayPoints/main.swift
@@ -89,11 +89,11 @@ func subtract (a:[Int], b:[Int]) -> ([Int]) {
     return [a[0] - b[0], a[1] - b[1]]
 }
 
-func add (a:Dictionary<String,Int>, b:Dictionary<String,Int>) -> (Dictionary<String,Int>) {
+func add (a:Dictionary<String,Double>, b:Dictionary<String,Double>) -> (Dictionary<String,Double>) {
     return ["x": a["x"]! + b["x"]!, "y": a["y"]! + b["y"]!]
 }
 
-func subtract (a:Dictionary<String,Int>, b:Dictionary<String,Int>) -> (Dictionary<String,Int>) {
+func subtract (a:Dictionary<String,Double>, b:Dictionary<String,Double>) -> (Dictionary<String,Double>) {
     return ["x": a["x"]! - b["x"]!, "y": a["y"]! - b["y"]!]
 }
